Fix shooting and passing
The ball object was a Mob which meant that it set velocity and overrode
the current forces applied every FixedUpdate. Remove the Mob component.

I also gave the player more angular drag (it was often spinning wildly)
and some linear drag (it was zero before). I'm not sure if the linear
will do anything since it's a Mob (see above).

Removed old hack code that was added for debugging.