Prevent rapid spinning of the player
Lock rotation around Z to prevent rapid spinning of the player after
colliding with walls. Not sure if this only occurred while holding the
ball, but that's when it was most noticeable.

Alternatively, I could solve this by giving them a large angular drag,
but that makes the problem less bad and I'd need to re-orient them back
to proper orientation.